requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
beautifulsoup4==4.12.2
bidict==0.22.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
discord==2.2.2
discord.py==2.2.2
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
frozenlist==1.3.3
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
multidict==6.0.4
ndg-httpsclient==0.5.1
Pillow==9.5.0
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
python-engineio==4.4.1
python-socketio==5.8.0
pytz==2023.3
pywatchman==1.4.1
redis==4.5.5
requests==2.28.2
rsa==4.9
schedule==1.2.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
yarl==1.9.1
zope.interface==6.0
